optimized trainer structure. support torch.profiler.
